feat: re-architecture of the sync onboarding
